Drop Scala 2.12 compatibility layer from //daml-lf/data
changelog_begin
changelog_end
